Split SearchResultsListView.vue into a component and a composable
