+++ frontend/src/components/search_results/SearchResultsListView.vue
@@ -1,33 +1,14 @@
 <script setup>
-import { ref, onMounted } from 'vue'
-import { useRouter } from 'vue-router'
-import axios from 'axios'
-import { checkLoginStatus } from '@/components/utils.js'
+import { onMounted } from 'vue'
+import { useSearchResultsList } from '@/composables/useSearchResultsList.js'
 
-const API_BASE_URL = import.meta.env.VITE_API_BASE_URL
 const emit = defineEmits(['message'])
-const router = useRouter()
-const samples = ref([])
-
-const fetchSearchResults = async () => {
-  try {
-    const response = await axios.get(`${API_BASE_URL}/list_search`)
-    samples.value = response.data
-  } catch (error) {
-    if (error.response && error.response.status === 404) {
-      emit('message', { type: 'danger', text: 'サンプルの取得に失敗しました。' })
-      router.replace({ name: 'NotFound' })
-    }
-  }
-}
+const { samples, loggedIn, fetchSearchResults } = useSearchResultsList(emit)
 
 onMounted(async () => {
-  const loggedIn = await checkLoginStatus(() => {
-    emit('message', { type: 'danger', text: 'ログインが必要です。' })
-    router.push('/')
-  })
-  if (!loggedIn) return
-  await fetchSearchResults()
+  if ( await loggedIn) {
+    await fetchSearchResults()
+  }
 })
 </script>
